Increase the limit for SCSI disk bus to 256
Kept the default number of disks to 26; while introducing 256 as
new limit for virtio-scsi devices. In the case of virtio-scsi;
libivirt uses one PCI storage controller per 7 disks. Depending on
the underlying CPU; the maximum number of PCI devices could be up
to 32 with 8 disks for each PCI device; value can be calculated by:

32x8=256

In the case of Rahti; a maximum number of pods which can be ran on
a single VM is 250. The limit can be higher if we want to be; but
that comes with consequences.